+++ ChangeLog.md
@@ -1,6 +1,6 @@
 # Revision history for th-abstraction
 
-## next -- ????.??.??
+## 0.7.0.0 -- ????.??.??
 * `normalizeCon` now takes a `Kind` argument representing the return kind of
   the parent data type. (This is sometimes necessary to determine which type
   variables in the data constructor are universal or existential, depending